Webアプリ開発2（3-2-5）指摘事項取込
