+++ web_application_2/server/functions/executeSql.js
@@ -1,0 +1,5 @@
+export default function executeSql (connection,sql) {
+    const statement = connection.prepareStatement(sql);
+    
+    return statement;  
+}
