+++ web_application_2/src/view/BookApp.vue
@@ -13,11 +13,11 @@
       @onClickDeleteIcon="onClickDeleteIcon"
      />
     <v-dialog v-model="isShowDialog" max-width="500px">
-     <form-template
-       :dialogBook="dialogBook"
-       @closeDialog="closeDialog"
-       @onClickSaveBtn="onClickSaveBtn"
-      />
+      <form-template
+        :dialogBook="dialogBook"
+        @closeDialog="closeDialog"
+        @onClickSaveBtn="onClickSaveBtn"
+       />
     </v-dialog>
     <v-dialog v-model="isShowDeleteDialog" max-width="500px">
      <v-card>
@@ -26,16 +26,6 @@
          <v-spacer />
          <v-btn @click="closeDeleteDialog">キャンセル</v-btn>
          <v-btn @click="onClickDeleteBtn">削除</v-btn>
-         <v-spacer />
-       </v-card-actions>
-     </v-card>
-    </v-dialog>
-    <v-dialog v-model="isShowErrorDialog" max-width="500px">
-     <v-card>
-       <v-card-title class="text-h5">未入力項目があります</v-card-title>
-       <v-card-actions>
-         <v-spacer />
-         <v-btn @click="closeErrorDialog">閉じる</v-btn>
          <v-spacer />
        </v-card-actions>
      </v-card>
@@ -68,7 +58,6 @@ export default {
       dialogBook: { title: '', category: '', purchase_date: '', buyer: '', review_content: '' },
       isShowDialog: false,
       isShowDeleteDialog: false,
-      isShowErrorDialog: false,
       searchTitle: '',
       searchCategory: '',
       isLoading: false
@@ -103,41 +92,28 @@ export default {
       this.isShowDialog = true
     },
     closeDialog () {
-      this.beforeDialogBook = {}
       this.isShowDialog = false
     },
     closeDeleteDialog () {
       this.isShowDeleteDialog = false
     },
-    closeErrorDialog () {
-      this.isShowErrorDialog = false
-    },
     async onClickSaveBtn (dialogBook) {
       this.dialogBook = dialogBook
-      if (this.dialogBook.title === undefined || this.dialogBook.title === '' ||
-          this.dialogBook.category === undefined || this.dialogBook.category === '' ||
-          this.dialogBook.purchase_date === undefined || this.dialogBook.purchase_date === '' ||
-          this.dialogBook.buyer === undefined || this.dialogBook.buyer === '' ||
-          this.dialogBook.review_content === undefined || this.dialogBook.review_content === ''
-      ) {
-        this.isShowErrorDialog = true
+      const google = window.google
+      if (Object.keys(this.beforeDialogBook).length === 0) {
+        await google.script.run.withSuccessHandler(function () {
+          alert('登録しました。')
+        }).withFailureHandler(function () {
+          alert('登録に失敗しました。')
+        }).insertRecord(this.dialogBook)
       } else {
-        const google = window.google
-        if (Object.keys(this.beforeDialogBook).length === 0) {
-          await google.script.run.withSuccessHandler(function () {
-            alert('登録しました。')
-          }).withFailureHandler(function () {
-            alert('登録に失敗しました。')
-          }).insertRecord(this.dialogBook)
-        } else {
-          await google.script.run.withSuccessHandler(function () {
-            alert('更新しました。')
-          }).withFailureHandler(function () {
-            alert('更新に失敗しました。')
-          }).updateRecord(this.beforeDialogBook, this.dialogBook)
-        }
-        this.closeDialog()
+        await google.script.run.withSuccessHandler(function () {
+          alert('更新しました。')
+        }).withFailureHandler(function () {
+          alert('更新に失敗しました。')
+        }).updateRecord(this.beforeDialogBook, this.dialogBook)
       }
+      this.closeDialog()
     },
     onClickDeleteBtn (deleteDialogBook) {
       // await google.script.run.withSuccessHandler(function () {
